Revise Scene implementation
Scene is now only responsible for drawing and keeping track of cameras.
It's now possible to remove elements from the render queue or changing
depths.
It also introduces the SceneObject class which is an extended Renderable
with render flags and depth.

VimCam was renamed to FreeCam and more keybindings were added.

A subsystem for Updatables is currently not available.

The old Scene implementation was moved to legacy/.