Fix access of entities created outside of EntityManager
Also remove move-assign add() function from EntityManager.

When an entity is created outside of EntityManager they don't get a
handle, which means, that components can't access the entity.
As a workaround an Entity* field was added that is used when the entity
was created outside.

The move-assign() add function was removed because it would require that
the handle and pointer field in all components get updated when the
entity is moved to the EntityManager.