Revise SpriteComponent
As long as there's no way of uniquely identifying a component in an
entity, there needs to be a workaround if two components depend on each
other.
SpriteComponent depends on an UpdateComponent to update the animation.
SpriteComponent now instantiates a corresponding UpdateComponent itself
without adding it to the entity.
Dirty hack, but works for now.