Allow using deque instead of vector in SlotMap
Deque has the advantage that pointers to elements aren't invalidated if
the container needs to reallocate.

EntityManager now uses a deque SlotMap, so that pointers to entities
don't get invalidated when the underlying container needs to reallocate.
Otherwise this would lead to segfaults and hard to fix bugs when
Transformables get a parent pointer and transformable components whose
parent pointer points to the entity's GroupTransform. When the
underlying vector reallocates, the parent pointer would get invalidated,
which results in segfaults.