Add CameraComponent and camera entity
Kinda works.
Creating a camera entity in the editor results in the editor cam and the
new cam being rendered.

It is also not possible to write a Transformable wrapper around Camera,
because the Camera could still be modified outside the Transformable
context, resulting in Transformable and Camera being out of sync.

Camera has to be based directly from Transformable to make this work.